fix/737: delete problem google calendar

+++ common/service/impl/EpGoogleCalenderServiceImpl.java
@@ -409,13 +409,17 @@ public class EpGoogleCalenderServiceImpl implements EpGoogleCalenderService {
 	public void deleteOutOfOfficeEvent(String eventId, String accessToken) {
 		Calendar service = createCalendarInstance(accessToken);
 		try {
-			service.events().delete(EpCommonConstants.CALENDAR_ID, eventId).execute();
-			log.info("GoogleCalendar: delete Event: {}", eventId);
+			Event event = service.events().get(EpCommonConstants.CALENDAR_ID, eventId).execute();
+			if (event != null) {
+				service.events().delete(EpCommonConstants.CALENDAR_ID, eventId).execute();
+				log.info("GoogleCalendar: deleted Event: {}", eventId);
 
-			calendarEventDao.deleteByEventId(eventId);
-		}
-		catch (Exception exception) {
-			log.error("GoogleCalendar: delete Event: {}", exception.getMessage(), exception);
+				calendarEventDao.deleteByEventId(eventId);
+			} else {
+				log.warn("GoogleCalendar: Event {} not found, nothing to delete", eventId);
+			}
+		} catch (Exception exception) {
+			log.error("GoogleCalendar: Error deleting Event {}: {}", eventId, exception.getMessage(), exception);
 			throw new ModuleException(EPCommonMessageConstant.EP_COMMON_UNABLE_TO_DELETE_GOOGLE_CALENDAR);
 		}
 	}
